# public/openapi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DebtResolve API",
        "contact": {
            "email": "jadiael@hotmail.com.br"
        },
        "license": {
            "name": "CC-BY-NC",
            "url": "https://creativecommons.org/licenses/by-nc/4.0/"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/installments": {
            "get": {
                "tags": [
                    "Installments"
                ],
                "summary": "List Installments",
                "description": "Retrieve a list of installments.",
                "operationId": "f50454fabde6c295db65732550a3dc6d",
                "responses": {
                    "200": {
                        "description": "Installments listed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Installments listed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "installments": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/installments/{installment}/installment": {
            "get": {
                "tags": [
                    "Installments"
                ],
                "summary": "Show Installment",
                "description": "Retrieve a specific installment by ID.",
                "operationId": "b9ff30783d7a46600f618d327775e17c",
                "parameters": [
                    {
                        "name": "installment",
                        "in": "path",
                        "description": "ID of the installment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Installment found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Installment found successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "installment": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Installment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Installment not found"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/installments/charge/{charge_id}/": {
            "get": {
                "tags": [
                    "Installments"
                ],
                "summary": "List installments for a specific charge",
                "description": "Retrieve a list of installments associated with a specific charge.",
                "operationId": "listInstallments",
                "parameters": [
                    {
                        "name": "charge",
                        "in": "path",
                        "description": "ID of the charge to retrieve installments",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Installments listed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Charge installment listing completed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "Installments": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Installment"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Charge not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Charge not found"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/installments/{installment}/charge/{charge}/accept-payment-approval-by-collector": {
            "post": {
                "tags": [
                    "Installments"
                ],
                "summary": "Accept Payment Approval by Collector",
                "description": "Accept payment approval for a specific installment by the collector.",
                "operationId": "26716fa4b31d16caa6485af45845d81c",
                "parameters": [
                    {
                        "name": "charge",
                        "in": "path",
                        "description": "ID of the charge",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "installment",
                        "in": "path",
                        "description": "ID of the installment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request body",
                    "required": false,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment marked as paid successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment marked as paid successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - You are not the collector of this charge and therefore cannot approve payments for this charge",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You are not the collector of this charge and therefore cannot approve payments for this charge"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Installment": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "description": "Value of the installment",
                        "type": "number",
                        "example": 100
                    },
                    "installment_number": {
                        "description": "Number of the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "due_date": {
                        "description": "Due date for payment",
                        "type": "string",
                        "format": "date",
                        "example": "2023-11-30"
                    },
                    "paid": {
                        "description": "Indicates if the installment is paid or not",
                        "type": "boolean",
                        "example": true
                    },
                    "user_id": {
                        "description": "ID of the user associated with the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "awaiting_approval": {
                        "description": "Indicates if the installment is awaiting approval",
                        "type": "boolean",
                        "example": false
                    },
                    "payment_proof": {
                        "description": "URL or path to the payment proof file",
                        "type": "string",
                        "example": "https://example.com/payment-proof.jpg"
                    },
                    "charge": {
                        "$ref": "#/components/schemas/Charge"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    },
...
}